third_party/WebKit/LayoutTests/fast/ime/inputmethodcontext-confirmcomposition.html

<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<input id="test">
<input id="test2">
<script>
description("This tests confirmComposition() of InputMethodContext.");

var test = document.getElementById('test');
var test2 = document.getElementById('test2');

// Checks .inputMethodContext returns an InputMethodContext object.
var context = test.inputMethodContext;
shouldBeTrue('context instanceof InputMethodContext');

var expectedText;

// Register event listeners.
test.addEventListener('compositionstart', function(event) {
    testPassed('compositionstart fired');
}, false);
test.addEventListener('compositionend', function(event) {
    testPassed('compositionend fired');
    shouldBeEqualToString('event.type', 'compositionend');
    shouldBeEqualToString('event.data', expectedText);
}, false);
test.addEventListener('textInput', function(event) {
    testPassed('textInput fired');
    shouldBeEqualToString('event.type', 'textInput');
    shouldBeEqualToString('event.data', expectedText);
}, false);

test.focus();

debug("Check if composition text 'abcd' is committed properly.");
expectedText = 'abcd';
textInputController.setComposition(expectedText);
context.confirmComposition();
shouldBeEqualToString('test.value', 'abcd');

debug("Check if no compositionend event will occur when composition is empty.");
context.confirmComposition();
shouldBeEqualToString('test.value', 'abcd');

debug("Check if composition text 'efgh' is committed properly, and appended to the input field's value.");
expectedText = 'efgh';
textInputController.setComposition(expectedText);
context.confirmComposition();
shouldBeEqualToString('test.value', 'abcdefgh');

debug("Check if confirmComposition on non-focused element doesn't affect the focused editor.");
var context2 = test2.inputMethodContext;
// Focus will still stay in the original field.
test.focus();
expectedText = 'ijkl';
textInputController.setComposition(expectedText);
// As composition for context2 is not on-going, no events will fire.
context2.confirmComposition();
// Cancel the composition.
expectedText = '';
textInputController.setComposition(expectedText);
shouldBeEqualToString('test.value', 'abcdefgh');
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
